# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Fuel Good API",
      "description": "API 測試頁面",
      "version": "0.0.1"
    },
    "servers": [
      {
        "url": "https://fuel-good.herokuapp.com/",
        "description": "Heroku"
      }
    ],
    "paths": {
      "/auth/testAPI": {
        "get": {
          "summary": "API連線測試",
          "description": "確認連線是否正常",
          "responses": {
            "200": {
              "description": "Message",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/google": {
        "get": {
            "tags": [
                "USER"
            ],
          "summary": "Google登入",
          "description": "透過Google Oauth登入",
          "responses": {
            "200": {
              "description": "null",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "null"
                  }
                }
              }
            }
          }
        }
      },    
      "/auth/login": {
        "post": {
          "tags": [
            "USER"
          ],
          "summary": "登入系統",
          "operationId": "loginUser",
          "requestBody":{
            "content":{
              "application/x-www-form-urlencoded":{
                "schema": {
                  "type": "object",
                  "properties":{
                    "email":{
                      "type":"string"
                    },
                    "password":{
                      "type":"string"
                    }
                  }
                }

              }
            }

        },"responses": {
          "200": {
            "description": "成功！",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied",
            "content": {}
          }
        }
        }
      },
      "/auth/signup": {
        "post": {
          "tags": [
            "USER"
          ],
          "summary": "用戶註冊",
          "operationId": "signupUser",
          "requestBody":{
            "content":{
              "application/x-www-form-urlencoded":{
                "schema": {
                  "type": "object",
                  "properties":{
                    "name":{
                      "type":"string"
                    },
                    "email":{
                      "type":"string"
                    },
                    "password":{
                      "type":"string"
                    }
                  }
                }

              }
            }

        },"responses": {
          "200": {
            "description": "成功！",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username/password/email supplied",
            "content": {}
          }
        }
        }
      },
      "/admin/user/all": {
        "get": {
          "tags": [
            "ADMIN"
          ],
          "summary": "列出所有已註冊用戶",
          "operationId": "findAllUser",
          "responses": {
            "200": {
              "description": "成功！"
            },
            "400": {
              "description": "連線問題",
              "content": {}
            }
          }
        }
      },
      "/admin/user/{name}": {
        "get": {
          "tags": [
            "ADMIN"
          ],
          "summary": "使用{name}找出特定用戶",
          "operationId": "findSpecificUser",
          "parameters": [
            {
              "in": "path",
              "name": "name",
              "description": "輸入使用者名稱",
              "required": true,
              "schema": {
                "type": "string",
                "properties": {
                  "name": {
                    "type": "string"
                    }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "成功！"
            },
            "400": {
              "description": "連線問題",
              "content": {}
            }
          }
        },
        "delete": {
          "tags": [
            "ADMIN"
          ],
          "summary": "使用{name}刪除特定用戶",
          "operationId": "deleteSpecificUser",
          "parameters": [
            {
              "in": "path",
              "name": "name",
              "description": "輸入使用者名稱",
              "required": true,
              "schema": {
                "type": "string",
                "properties": {
                  "name": {
                    "type": "string"
                    }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "成功！"
            },
            "400": {
              "description": "連線問題",
              "content": {}
            }
          }
        }
      },
      "/crawler/price/recent/": {
        "get": {
          "tags": [
            "PRICE"
          ],
          "summary": "列出當期油價",
          "operationId": "findRecentPrice",
          "responses": {
            "200": {
              "description": "成功！"
            },
            "400": {
              "description": "連線問題",
              "content": {}
            }
          }
        }
      },
      "/crawler/price/history/all": {
        "get": {
          "tags": [
            "PRICE"
          ],
          "summary": "列出所有油價",
          "operationId": "findAllPrice",
          "responses": {
            "200": {
              "description": "成功！"
            },
            "400": {
              "description": "連線問題",
              "content": {}
            }
          }
        }
      },
      "/crawler/price/history/{gasoline}": {
        "get": {
          "tags": [
            "PRICE"
          ],
          "summary": "列出特定汽油之所有油價歷史",
          "operationId": "findOnePrice",
          "parameters": [
            {
              "in": "path",
              "name": "gasoline",
              "description": "輸入汽油種類",
              "required": true,
              "schema": {
                "type": "string",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }  
              }
            }

          ],
          "responses": {
            "200": {
              "description": "成功！"
            },
            "400": {
              "description": "連線問題",
              "content": {}
            }
          }
        }
      },
      "/order/add": {
        "post": {
          "tags": [
            "ORDER"
          ],
          "summary": "新增訂單",
          "operationId": "addOrder",
          "requestBody":{
            "content":{
              "application/x-www-form-urlencoded":{
                "schema": {
                  "type": "object",
                  "properties":{
                    "email":{
                      "type":"string"
                    },
                    "adding92":{
                      "type":"number"
                    },
                    "adding95":{
                      "type":"number"
                    },
                    "adding98":{
                      "type":"number"
                    },
                    "addingDiesel":{
                      "type":"number"
                    }
                  }
                }

              }
            }

        },"responses": {
          "200": {
            "description": "新增訂單成功！",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "訂單新增失敗",
            "content": {}
          }
        }
        }
      },
      "/order/deduct": {
        "patch": {
          "tags": [
            "ORDER"
          ],
          "summary": "取用汽油扣除存量",
          "operationId": "deductQty",
          "requestBody":{
            "content":{
              "application/x-www-form-urlencoded":{
                "schema": {
                  "type": "object",
                  "properties":{
                    "email":{
                      "type":"string"
                    },
                    "deducted92":{
                      "type":"number"
                    },
                    "deducted95":{
                      "type":"number"
                    },
                    "deducted98":{
                      "type":"number"
                    },
                    "deductedDiesel":{
                      "type":"number"
                    }
                  }
                }

              }
            }

        },"responses": {
          "200": {
            "description": "庫存取用成功！",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "庫存取用失敗",
            "content": {}
          }
        }
        }
      },
      "/order/all": {
        "get": {
          "tags": [
            "ORDER"
          ],
          "summary": "列出所有訂單",
          "operationId": "findAllOrder",
          "responses": {
            "200": {
              "description": "成功列出所有訂單"
            },
            "400": {
              "description": "發生問題",
              "content": {}
            }
          }
        }
      }

    }
}